+++ src/main/java/de/catma/rbac/RBACSubject.java
@@ -3,7 +3,7 @@ package de.catma.rbac;
 import de.catma.user.User;
 
 /**
- * Subject in the RBAC system. Usually this is implemented in {@link User} with the {@link #User.getUserId()} method
+ * Subject in the RBAC system. Usually this is implemented in {@link User} with the {@link User#getUserId()} method
  *
  * @author db
  *
@@ -14,5 +14,5 @@ public interface RBACSubject {
 	 * Retrieves the identifier of a subject
 	 * @return the identifier
 	 */
-	Integer getUserId();
+	Long getUserId();
 }
