+++ src/main/java/de/catma/ui/module/project/ProjectInvitationHandler.java
@@ -23,9 +23,9 @@ import de.catma.util.ColorConverter;
 public class ProjectInvitationHandler extends UIMessageListener<InvitationRequestMessage> {
 	
 	private final Map<String, DocumentResource> documentResourceByUuid;
-	private final Map<Integer, String> userId2Color = new HashMap<>();
+	private final Map<Long, String> userId2Color = new HashMap<>();
 	private final Set<String> colors = ColorConverter.getColorNames();
-	private final Set<Integer> assignedUsers = new HashSet<>();
+	private final Set<Long> assignedUsers = new HashSet<>();
 	private final ProjectInvitation projectInvitation;
 	private final Project project;
 	private final ITopic<JoinedProjectMessage> joinedTopic;
@@ -53,7 +53,7 @@ public class ProjectInvitationHandler extends UIMessageListener<InvitationReques
 	public void uiOnMessage(Message<InvitationRequestMessage> message) {
 		if (message.getMessageObject().getCode() == projectInvitation.getKey()) {
 			try {
-				Integer userId = Integer.valueOf(message.getMessageObject().getUserId());
+				Long userId = Long.valueOf(message.getMessageObject().getUserId());
 				
 				if (!assignedUsers.contains(userId)) {
